Added Filters and added some refactors
Added the filters categarory;
Added gaussian_filter.py

Refactored the raw code into the src folder for better integration in future.
