+++ Arrays-searching/src/binary_search.py
