+++ Json Parser/src/Java/jsonHelperArray.java
